refactor: Added helper tagHasFlogEntryDate to useTags, passed through useDropboxFlogs, useFlogSource
